# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Core Accounting MVP API
  version: 1.2.0
  description: 'مواصفات REST API للنواة المحاسبية القابلة للتوسع.


    مصادر التصميم: ERD v1.2، قاموس البيانات v1.2، مصفوفة الصلاحيات وقواعد الترحيل v1.1، وحالات الاستخدام وتدفقات العمل v1.0، مع تعديلات ADR-002 v1.3.


    قرارات تنفيذية مثبتة في هذه النسخة:

    - الشركة الحالية تُثبت في الجلسة عبر /auth/context، ولا يُعتمد companyId الوارد من العميل في العمليات التشغيلية.

    - رقم المستند يُحجز عند إنشاء المسودة مع قبول فجوات الأرقام الموثقة.

    - منشئ القيد اليدوي لا يرحله افتراضياً، ويمكن جعل ذلك إعداداً للشركات الصغيرة.

    - واجهة التخصيصات مدعومة في API، ولو أُجلت واجهتها في الإصدار الأول.

    - مدة الاحتفاظ بمفتاح Idempotency هي 24 ساعة افتراضياً وقابلة للضبط.

    - أقصى نطاق تفاعلي للتقارير 366 يوماً.


    المبالغ وأسعار الصرف والمعرفات الكبيرة تُنقل كنصوص للحفاظ على الدقة. الحقول المشفرة وBlind Indexes لا تظهر في الاستجابات
    العامة.


    تغييرات v1.1:

    - اعتماد counter_account_id رسمياً في سندات القبض والصرف العامة.

    - إضافة GET /dashboard/summary ببيانات مشتقة وصلاحيات حقلية.

    - لا تُخزن مؤشرات لوحة المتابعة كمصدر حقيقة مستقل.'
  contact:
    name: Core Accounting MVP Technical Team
  license:
    name: Internal Use
    identifier: LicenseRef-Internal
servers:
- url: /api/v1
  description: مسار نسبي يعمل على نفس أصل الواجهة في جميع البيئات
...
security:
- SessionCookie: []
paths:
...
  /auth/csrf:
    get:
      tags:
      - Authentication
      summary: إنشاء أو تحديث رمز CSRF
      operationId: getCsrfToken
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfTokenResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: ينشئ جلسة تمهيدية عند الحاجة ويعيد رمز CSRF. تُرسل Cookie الجلسة بخصائص HttpOnly وSameSite، وSecure في
        الإنتاج.
      security: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
  /auth/login:
    post:
      tags:
      - Authentication
      summary: تسجيل الدخول
      operationId: login
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: رسالة الخطأ لا تكشف ما إذا كان البريد موجوداً. بعد خمس محاولات فاشلة يُقفل الحساب 15 دقيقة وفق السياسة
        المعتمدة.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      security:
      - SessionCookie: []
        CsrfToken: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
...
  /auth/companies:
    get:
      tags:
      - Authentication
      summary: الشركات المصرح بها للمستخدم
      operationId: listAccessibleCompanies
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyAccessList'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      x-permission: companies.view
      security:
      - SessionCookie: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
  /auth/context:
    post:
      tags:
      - Authentication
      summary: اختيار الشركة الحالية
      operationId: selectCompanyContext
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: يثبت companyId في الجلسة. جميع المسارات التشغيلية ال تقبل companyId من جسم الطلب كمصدر موثوق.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectCompanyRequest'
      x-permission: companies.view
      security:
      - SessionCookie: []
        CsrfToken: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
  /auth/logout:
    post:
      tags:
      - Authentication
      summary: إنهاء الجلسة الحالية
      operationId: logout
      responses:
        '204':
          description: تم إلغاء الجلسة وإزالة Cookie
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
      - SessionCookie: []
        CsrfToken: []
      parameters:
      - $ref: '#/components/parameters/RequestId'
  /auth/sessions:
    get:
      tags:
      - Authentication
      summary: عرض جلسات المستخدم
      operationId: listSessions
      responses:
        '200':
          description: قائمة النتائج
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      x-permission: auth.sessions.view
      security:
      - SessionCookie: []
  /auth/sessions/{sessionId}/revoke:
    post:
      tags:
      - Authentication
      summary: إلغاء جلسة
      operationId: revokeSession
      responses:
        '204':
          description: ألغيت الجلسة أو كانت ملغاة مسبقاً
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: sessionId
        in: path
        required: true
        description: معرف الجلسة
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      x-permission: auth.sessions.revoke
      security:
      - SessionCookie: []
        CsrfToken: []
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: معرف جلسة عشوائي؛ القيمة الخام لا تُخزن في قاعدة البيانات.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: رمز Synchronizer مرتبط بالجلسة ومطلوب لكل طلب يغير الحالة.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: خطأ في التنسيق أو الحقول
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: الجلسة مفقودة أو منتهية أو ملغاة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: لا توجد صلاحية أو سياق الشركة غير مصرح به
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: الكيان غير موجود داخل الشركة الحالية؛ لا يكشف وجوده في شركة أخرى
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: تعارض إصدار أو تزامن أو تكرار
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    BusinessRuleViolation:
      description: انتهاك قاعدة أعمال، مثل فترة مغلقة أو قيد غير متوازن
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: تم تجاوز معدل الطلبات
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: خطأ داخلي؛ يعاد requestId فقط دون تفاصيل حساسة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    PaginationMeta:
      type: object
      required:
      - page
      - pageSize
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Problem:
      type: object
      required:
      - code
      - messageAr
      - requestId
      properties:
        code:
          type: string
          enum:
          - VALIDATION_ERROR
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - BUSINESS_RULE_VIOLATION
          - RATE_LIMITED
          - INTERNAL_ERROR
        messageAr:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: object
          additionalProperties: true
          description: تفاصيل آمنة فقط؛ لا تتضمن SQL أو Stack Trace أو قيماً حساسة.
...
    CsrfTokenResponse:
      type: object
      required:
      - csrfToken
      properties:
        csrfToken:
          type: string
          minLength: 32
        expiresAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
    LoginRequest:
      type: object
      additionalProperties: false
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 200
    SelectCompanyRequest:
      type: object
      additionalProperties: false
      required:
      - companyId
      properties:
        companyId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
    CompanyAccess:
      type: object
      required:
      - companyId
      - code
      - legalNameAr
      - roles
      - permissions
      properties:
        companyId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        code:
          type: string
        legalNameAr:
          type: string
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    CompanyAccessList:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CompanyAccess'
    AuthenticatedUser:
      type: object
      required:
      - id
      - email
      - nameAr
      - status
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        email:
          type: string
          format: email
        nameAr:
          type: string
        nameEn:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - ACTIVE
          - LOCKED
          - DISABLED
    AuthSessionResponse:
      type: object
      required:
      - user
      - csrfToken
      - companies
      properties:
        user:
          $ref: '#/components/schemas/AuthenticatedUser'
        csrfToken:
          type: string
        selectedCompany:
          oneOf:
          - $ref: '#/components/schemas/CompanyAccess'
          - type: 'null'
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanyAccess'
        sessionExpiresAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
    Session:
      type: object
      required:
      - id
      - createdAt
      - lastActivityAt
      - expiresAt
      - current
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        createdAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
        lastActivityAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
        expiresAt:
          type: string
          format: date-time
          example: '2026-07-30T06:00:00.000Z'
        revokedAt:
          oneOf:
          - type: string
            format: date-time
            example: '2026-07-30T06:00:00.000Z'
          - type: 'null'
        current:
          type: boolean
        ipMasked:
          type:
          - string
          - 'null'
        userAgentSummary:
          type:
          - string
          - 'null'
...